+++ pkg/controllers/gitrepository/gitea.go
@@ -3,6 +3,7 @@ package gitrepository
 import (
 	"context"
 	"fmt"
+	"github.com/cnoe-io/idpbuilder/pkg/util/files"
 	"os"
 	"path/filepath"
 
@@ -136,7 +137,7 @@ func writeRepoContents(repo *v1alpha1.GitRepository, dstPath string, config v1al
 		return nil
 	}
 
-	err := util.CopyDirectory(repo.Spec.Source.Path, dstPath)
+	err := files.CopyDirectory(repo.Spec.Source.Path, dstPath)
 	if err != nil {
 		return fmt.Errorf("copying files: %w", err)
 	}
