+++ pkg/util/fs/fs_test.go
@@ -1,4 +1,4 @@
-package util
+package fs
 
 import (
 	"fmt"
